+++ maia/algo/part/localize.py
@@ -1,4 +1,5 @@
 import numpy as np
+from mpi4py import MPI
 
 import Pypdm.Pypdm as PDM
 
@@ -12,8 +13,8 @@ from maia.utils                  import vstride as vs
 
 from maia.factory.dist_from_part import get_parts_per_blocks
 
-from .point_cloud_utils  import get_shifted_point_clouds
-from .connectivity_utils import cell_vtx_connectivity_elts, PDM_connectivity_transpose
+from .point_cloud_utils  import get_point_cloud
+from .connectivity_utils import cell_vtx_connectivity, PDM_connectivity_transpose
 
 def _get_part_data_ngon(part_zone):
   dim = PT.Zone.CellDimension(part_zone)
@@ -57,11 +58,12 @@ def _get_part_data_ngon(part_zone):
 
 
 def _get_part_data_elts(part_zone):
-  cx, cy, cz = PT.Zone.coordinates(part_zone)
-  vtx_coords = np_utils.interweave_arrays([cx,cy,cz])
+  # Actually works for elt of S meshes, for which we rebuild cell_vtx connectivity
+  coords = [c.reshape(-1, order='F') for c in PT.Zone.coordinates(part_zone)]
+  vtx_coords = np_utils.interweave_arrays(coords)
 
   dim = PT.Zone.CellDimension(part_zone)
-  cell_vtx = cell_vtx_connectivity_elts(part_zone, dim)
+  cell_vtx = cell_vtx_connectivity(part_zone, dim)
 
   vtx_ln_to_gn, _, _, cell_ln_to_gn = te_utils.get_entities_numbering(part_zone)
 
@@ -72,7 +74,7 @@ def _mesh_location(src_parts, tgt_clouds, comm, reverse=False, loc_tolerance=1E-
   """ Wrapper of PDM mesh location
   For now, only 1 domain is supported so we expect source parts and target clouds
   as flat lists :
-  Parts are tuple dim, elt_kind, part_data 
+  Parts are tuple (dim, elt_kind, part_data)
    where dim = 2 or 3, elt_kind = 'Poly' or 'Element' and part_data stores the arrays
    expected by paradigm
   Cloud are tuple (coords, lngn)
@@ -132,45 +134,45 @@ def _mesh_location(src_parts, tgt_clouds, comm, reverse=False, loc_tolerance=1E-
   else:
     return all_target_data
 
-def _localize_points(src_parts_per_dom, tgt_parts_per_dom, location, comm, \
-    reverse=False, loc_tolerance=1E-6):
+def _mdom_mesh_location(src_parts_per_dom, tgt_clouds_per_dom, comm, reverse=False, loc_tolerance=1E-6):
   """
+  Wraps _mesh_location with multidomain support (with shifts)
+  Input are similar to _mesh_location, but with nested lists by domains
+  Input data must not be shifted, it will be done by this function
   """
-  locs = {'NGon2D' :{'Cell':2, 'Vtx':5},
-          'NGon3D' :{'Cell':2, 'Face':5, 'Vtx':7},
-          'Element':{'Cell':2, 'Vtx':4}}
-  n_dom_src = len(src_parts_per_dom)
-  n_dom_tgt = len(tgt_parts_per_dom)
+  
+  n_part_per_dom_src = [len(parts) for parts in src_parts_per_dom ]
+  n_part_per_dom_tgt = [len(parts) for parts in tgt_clouds_per_dom]
 
-  n_part_per_dom_src = [len(parts) for parts in src_parts_per_dom]
-  n_part_per_dom_tgt = [len(parts) for parts in tgt_parts_per_dom]
-  n_part_src = sum(n_part_per_dom_src)
-  n_part_tgt = sum(n_part_per_dom_tgt)
+  # Shift target data first; we dont do it inplace since src and target data
+  # may share the same memory
+  tgt_offset = np.zeros(len(tgt_clouds_per_dom)+1, dtype=pdm_gnum_dtype)
+  for i_domain, clouds in enumerate(tgt_clouds_per_dom):
+    # Compute global offsets for this domain
+    dom_max = par_utils.arrays_max([cloud[1] for cloud in clouds], comm)
+    tgt_offset[i_domain+1] = tgt_offset[i_domain] + dom_max
+    # Shift source arrays (copy)
+    tgt_clouds_per_dom[i_domain] = [(c[0], c[1] + tgt_offset[i_domain]) for c in clouds]
 
-  # > Register source
-  connectivity_t = None
-  src_parts = []
-  for i_domain, src_part_zones in enumerate(src_parts_per_dom):
-    # TODO : use cell_vtx_connectivity_ngon to transform ngon into nodal ?
+  tgt_clouds = py_utils.to_flat_list(tgt_clouds_per_dom)
+  src_parts  = py_utils.to_flat_list(src_parts_per_dom)
 
-    src_parts_domain = list()
-    for src_part in src_part_zones:
-      dim = PT.Zone.CellDimension(src_part)
-      if PT.Zone.has_ngon_elements(src_part):
-        if connectivity_t=='Element':
-          raise NotImplementedError("Source mesh must have NGon or Element connectivity but not both.")
-        connectivity_t = f'NGon{dim}D'
-        src_parts_domain.append((dim, 'Poly', _get_part_data_ngon(src_part)))
-      else:
-        if connectivity_t is not None and 'NGON' in connectivity_t:
-          raise NotImplementedError("Source mesh must have NGon or Element connectivity but not both.")
-        connectivity_t = 'Element'
-        src_parts_domain.append((dim, 'Element', _get_part_data_elts(src_part)))
-    src_parts.append(src_parts_domain)
+  # Now shift source data. First we need to eetrieve loc, which should be the same for each partition.
+  kind = ''
+  if len(src_parts) > 0:
+    dim, kind = src_parts[0][:2]
+    if kind == 'Poly':
+      kind += str(dim)
+  kind = comm.allreduce(kind, MPI.MAX) # To let empty procs know the data
 
-  locs = locs[connectivity_t]
-  src_offsets = {loc : np.zeros(n_dom_src+1, dtype=pdm_gnum_dtype) for loc in locs}
-  for i_domain, src_parts_domain in enumerate(src_parts):
+  locs = {'Poly2'   : {'Cell':2, 'Vtx':5},           
+          'Poly3'   : {'Cell':2, 'Face':5, 'Vtx':7},
+          'Element' : {'Cell':2, 'Vtx':4},         
+         }[kind]
+
+  # Effective shift
+  src_offsets = {loc : np.zeros(len(src_parts_per_dom)+1, dtype=pdm_gnum_dtype) for loc in locs}
+  for i_domain, src_parts_domain in enumerate(src_parts_per_dom):
     # Compute global offsets for this domain
     for loc, array_idx in locs.items():
       dom_max = par_utils.arrays_max([src_part[2][array_idx] for src_part in src_parts_domain], comm)
@@ -179,16 +181,11 @@ def _localize_points(src_parts_per_dom, tgt_parts_per_dom, location, comm, \
     for src_part in src_parts_domain:
       for loc, array_idx in locs.items():
         src_part[2][array_idx] += src_offsets[loc][i_domain]
-  
-  src_parts = py_utils.to_flat_list(src_parts)
-
-  tgt_offset, tgt_clouds = get_shifted_point_clouds(tgt_parts_per_dom, location, comm)
-  tgt_clouds = py_utils.to_flat_list(tgt_clouds)
 
   result = _mesh_location(src_parts, tgt_clouds, comm, reverse, loc_tolerance)
 
   # Shift back source data
-  for i_domain, src_parts_domain in enumerate(py_utils.to_nested_list(src_parts, n_part_per_dom_src)):
+  for i_domain, src_parts_domain in enumerate(src_parts_per_dom):
     for src_part in src_parts_domain:
       for loc, array_idx in locs.items():
         src_part[2][array_idx] -= src_offsets[loc][i_domain]
@@ -213,32 +210,49 @@ def _localize_points(src_parts_per_dom, tgt_parts_per_dom, location, comm, \
   else:
     return py_utils.to_nested_list(result, n_part_per_dom_tgt)
 
+def _collect_source(src_parts_per_dom):
+  connectivity_t = None
+  src_parts = []
+  for src_part_zones in src_parts_per_dom:
+
+    src_parts_domain = list()
+    for src_part in src_part_zones:
+      dim = PT.Zone.CellDimension(src_part)
+      if PT.Zone.has_ngon_elements(src_part):
+        if connectivity_t=='Element':
+          raise NotImplementedError("Source mesh must have NGon or Element connectivity but not both.")
+        connectivity_t = f'NGon{dim}D'
+        src_parts_domain.append((dim, 'Poly', _get_part_data_ngon(src_part)))
+      else:
+        if connectivity_t is not None and 'NGON' in connectivity_t:
+          raise NotImplementedError("Source mesh must have NGon or Element connectivity but not both.")
+        connectivity_t = 'Element'
+        src_parts_domain.append((dim, 'Element', _get_part_data_elts(src_part)))
+    src_parts.append(src_parts_domain)
+
+  return src_parts
+
+def _collect_target(tgt_parts_per_dom, location):
+  return [[get_point_cloud(part, location) for part in tgt_parts] \
+          for tgt_parts in tgt_parts_per_dom]
+
+
+def _localize_points(src_parts_per_dom, tgt_parts_per_dom, location, comm, \
+    reverse=False, loc_tolerance=1E-6):
+  """ Intermediate API who do not place output in tree.
+  Inputs are list of size n_domain_src (resp. tgt) containing partitioned zones (resp. clouds)
+  for each domain 
+  """
+  src_parts  = _collect_source(src_parts_per_dom)
+  tgt_clouds = _collect_target(tgt_parts_per_dom, location)
+
+  return _mdom_mesh_location(src_parts, tgt_clouds, comm, reverse, loc_tolerance)
+
+
+
 def localize_points(src_tree, tgt_tree, location, comm, **options):
-  """Localize points between two partitioned trees.
-
-  For all the points of the target tree matching the given location,
-  search the cell of the source tree in which it is enclosed.
-  The result, i.e. the gnum & domain number of the source cell (or -1 if the point is not localized),
-  are stored in a ``DiscreteData_t`` container called "Localization" on the target zones.
-
-  Source tree must be unstructured.
-
-  Localization can be parametred thought the options kwargs:
-
-  - ``loc_tolerance`` (default = 1E-6) -- Geometric tolerance for the method.
-
-  Args:
-    src_tree (CGNSTree): Source tree, partitionned. Only unstructured connectivities are managed.
-    tgt_tree (CGNSTree): Target tree, partitionned.
-    location ({'CellCenter', 'Vertex'}) : Target points to localize
-    comm       (MPIComm): MPI communicator
-    **options: Additional options related to location strategy
-
-  Example:
-      .. literalinclude:: snippets/test_algo.py
-        :start-after: #localize_points@start
-        :end-before: #localize_points@end
-        :dedent: 2
+  """
+  Partitionned implementation of maia.algo.localize_points
   """
   _src_parts_per_dom = get_parts_per_blocks(src_tree, comm)
   src_parts_per_dom = list(_src_parts_per_dom.values())
@@ -257,7 +271,18 @@ def localize_points(src_tree, tgt_tree, location, comm, **options):
       src_dom  = -np.ones(n_tgts, dtype=np.int32)
       src_gnum[data['located_ids']] = data['location']
       src_dom [data['located_ids']] = data['domain']
+      # For structured meshes, reshape result
+      if PT.Zone.Type(tgt_part) == 'Structured':
+        if n_tgts == PT.Zone.n_vtx(tgt_part):
+          shape = PT.Zone.VertexSize(tgt_part)
+        elif n_tgts == PT.Zone.n_cell(tgt_part):
+          shape = PT.Zone.CellSize(tgt_part)
+        else:
+          raise RuntimeError("Unable to detect target location")
+        src_gnum = src_gnum.reshape(shape, order='F')
+        src_dom  = src_dom.reshape(shape, order='F')
+
       PT.new_DataArray("SrcId", src_gnum, parent=sol)
       PT.new_DataArray("DomId", src_dom,  parent=sol)
-      PT.new_node("DomainList", "Descriptor_t", dom_list, parent=sol)
+      PT.new_Descriptor("DomainList", dom_list, parent=sol)